--- src/Services/Identify/Identify.BusinessLogic/DTOs/RequestDTOs/RefreshTokenDto.cs
@@ -1,3 +1,0 @@
-﻿namespace Identify.BusinessLogic.DTOs.RequestDTOs;
-
-public sealed record RefreshTokenDto(string AccessToken, string RefreshToken);
