--- src/Services/Identify/Identify.BusinessLogic/DTOs/RequestDTOs/RegisterUserDto.cs
@@ -1,6 +1,0 @@
-﻿namespace Identify.BusinessLogic.DTOs.RequestDTOs;
-
-public sealed record RegisterUserDto(
-    string UserName,
-    string Email,
-    string Password);
