--- src/Services/Identify/Identify.BusinessLogic/DTOs/RequestDTOs/UpdateUserDto.cs
@@ -1,6 +1,0 @@
-﻿namespace Identify.BusinessLogic.DTOs.RequestDTOs;
-
-public sealed record UpdateUserDto(
-    int Id,
-    string UserName,
-    string Email);
