--- src/Services/Identify/Identify.BusinessLogic/DTOs/RequestDTOs/UserCredentialsDto.cs
@@ -1,3 +1,0 @@
-﻿namespace Identify.BusinessLogic.DTOs.RequestDTOs;
-
-public sealed record UserCredentialsDto(string UserName, string Password);
