--- src/Services/Identify/Identify.BusinessLogic/DTOs/ResponseDTOs/TokenDto.cs
@@ -1,6 +1,0 @@
-﻿namespace Identify.BusinessLogic.DTOs.ResponseDTOs;
-
-public sealed record TokenDto(
-    string AccessToken,
-    DateTime Expires,
-    string RefreshToken);
