--- src/Services/Identify/Identify.BusinessLogic/DTOs/ResponseDTOs/UserDto.cs
@@ -1,9 +1,0 @@
-﻿namespace Identify.BusinessLogic.DTOs.ResponseDTOs;
-
-public sealed class UserDto
-{
-    public int Id { get; set; }
-    public string Email { get; set; }
-    public string UserName { get; set; }
-    public IList<string> Roles { get; set; } 
-}
