--- src/Services/Identify/Identify.DataAccess/Entities/User.cs
@@ -1,14 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-using Shared;
-
-namespace Identify.DataAccess.Entities;
-
-public class User : IdentityUser<int>, IBaseEntity
-{
-    public string? ImageUrl { get; set; }
-    public string? RefreshToken { get; set; }
-    public DateTime? RefreshTokenExpiryTime { get; set; }
-    public DateTimeOffset DateCreated { get; set; }
-    public DateTimeOffset? DateUpdated { get; set; }
-    public DateTimeOffset? DateDeleted { get; set; }
-}
