--- src/Services/Identify/Identify.DataAccess/Enums/Roles.cs
@@ -1,8 +1,0 @@
-﻿namespace Identify.DataAccess.Enums;
-
-public enum Roles
-{
-    Client,
-    Admin,
-    SuperAdmin
-}
