--- src/Services/Identify/Identify.DataAccess/Identify.DataAccess.csproj
@@ -1,27 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <RootNamespace>Identify.DataAccess</RootNamespace>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <Folder Include="Contexts\" />
-        <Folder Include="Entities\" />
-        <Folder Include="EntityConfigurations\" />
-        <Folder Include="Errors\" />
-        <Folder Include="Interfaces\" />
-        <Folder Include="Migrations\" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\..\..\Shared\Shared.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="8.0.3" />
-    </ItemGroup>
-
-</Project>
