+++ src/Services/Identity/Identity.BusinessLogic/DTOs/RequestDTOs/RefreshTokenDto.cs
@@ -1,0 +1,3 @@
+﻿namespace Identity.BusinessLogic.DTOs.RequestDTOs;
+
+public sealed record RefreshTokenDto(string AccessToken, string RefreshToken);
