+++ src/Services/Identity/Identity.BusinessLogic/DTOs/RequestDTOs/RegisterUserDto.cs
@@ -1,0 +1,6 @@
+﻿namespace Identity.BusinessLogic.DTOs.RequestDTOs;
+
+public sealed record RegisterUserDto(
+    string UserName,
+    string Email,
+    string Password);
