+++ src/Services/Identity/Identity.BusinessLogic/DTOs/RequestDTOs/UpdateUserDto.cs
@@ -1,0 +1,6 @@
+﻿namespace Identity.BusinessLogic.DTOs.RequestDTOs;
+
+public sealed record UpdateUserDto(
+    int Id,
+    string UserName,
+    string Email);
