+++ src/Services/Identity/Identity.BusinessLogic/DTOs/RequestDTOs/UserCredentialsDto.cs
@@ -1,0 +1,3 @@
+﻿namespace Identity.BusinessLogic.DTOs.RequestDTOs;
+
+public sealed record UserCredentialsDto(string UserName, string Password);
