+++ src/Services/Identity/Identity.BusinessLogic/DTOs/ResponseDTOs/TokenDto.cs
@@ -1,0 +1,6 @@
+﻿namespace Identity.BusinessLogic.DTOs.ResponseDTOs;
+
+public sealed record TokenDto(
+    string AccessToken,
+    DateTime Expires,
+    string RefreshToken);
