+++ src/Services/Identity/Identity.BusinessLogic/DTOs/ResponseDTOs/UserDto.cs
@@ -1,0 +1,9 @@
+﻿namespace Identity.BusinessLogic.DTOs.ResponseDTOs;
+
+public sealed class UserDto
+{
+    public int Id { get; set; }
+    public string Email { get; set; }
+    public string UserName { get; set; }
+    public IList<string> Roles { get; set; } 
+}
