+++ src/Services/Identity/Identity.BusinessLogic/Models/JwtOptions.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.IdentityModel.Tokens;
+
+namespace Identity.BusinessLogic.Models;
+
+public class JwtOptions
+{
+    public string Issuer { get; set; }
+    public string Key { get; set; }
+    public int TokenValidityInMinutes { get; set; }
+    public int RefreshTokenValidityInDays { get; set; }
+    public TokenValidationParameters ValidationParameters { get; set; }   
+}
