+++ src/Services/Identity/Identity.DataAccess/Entities/Role.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.AspNetCore.Identity;
+using Shared;
+
+namespace Identity.DataAccess.Entities;
+
+public class Role: IdentityRole<int>, IBaseEntity
+{
+    public DateTimeOffset DateCreated { get; set; }
+    public DateTimeOffset? DateUpdated { get; set; }
+    public DateTimeOffset? DateDeleted { get; set; }
+}
