+++ src/Services/Identity/Identity.DataAccess/Enums/Roles.cs
@@ -1,0 +1,8 @@
+﻿namespace Identity.DataAccess.Enums;
+
+public enum Roles
+{
+    Client,
+    Admin,
+    SuperAdmin
+}
